Miscellany

Here we show a few small functionalities that didn't fit anywhere else.

StructuredLight.overlap — Function

overlap(ψ₁,ψ₂,xs,ys)

Calculate ∫ ψ₁ conj(ψ₂) dx dy.

xs and ys are the grids over which ψ₁ and ψ₂ are calculated.

overlap(ψ₁::AbstractArray{T1,3},ψ₂::AbstractArray{T2,3},xs,ys) where {T1,T2}

Calculate ∫ ψ₁[:,:,j] conj(ψ₂[:,:,j]) dx dy, where j runs over all indices of the third dimension.

The output is a vector.

xs and ys are the grids over which ψ₁ and ψ₂ are calculated.

using StructuredLight, CairoMakie

rs = LinRange(-5,5,256) 
ψ₁ = hg(rs,rs)
ψ₂ = hg(rs,rs,m=1)

overlap(ψ₁,ψ₁,rs,rs) #The modes are normalized
overlap(ψ₁,ψ₂,rs,rs) #Modes with different indices are orthogonal

zs = LinRange(0,.5,8)
ψ₁s = free_propagation(ψ₁,rs,rs,zs)
ψ₂s = free_propagation(ψ₂,rs,rs,zs)

# Free propagation is an unitary transformation
# Therefore, it preserves the overlap
overlap(ψ₁s,ψ₁s,rs,rs), overlap(ψ₁s,ψ₂s,rs,rs)