Propagation

The heavy lifting in this package is done by the propagation methods. The basic principle is that you give an initial profile and the grid over which it was calculated, and we give back to you the propagated beam. For now, we have included propagation in free space (free_propagation) and in Kerr Media (kerr_propagation). The implementation of propagation in media with quadratic nonlinearity is on the horizon.

free_propagation

StructuredLight.free_propagation — Function

free_propagation(ψ₀, xs, ys, z::Number [, scaling]; k=1)
free_propagation(ψ₀,xs,ys,z::AbstractArray [, scaling]; k=1)

Propagate an inital profile ψ₀.

The propagation is the solution of ∇² ψ + 2ik ∂_z ψ = 0 at distance z under the initial condition ψ₀.

xs and ys are the grids over which ψ₀ is calculated.

If z is an AbstractArray, the output is a 3D array representing the solution at every element of z.

The output at a distance z[n] is calculated on a scalled grid defined by scaling[n] * xs and scaling[n] * ys.

k is the wavenumber.

Examples:

#The simplest usage would be the following:
using StructuredLight

rs = LinRange(-6,6,256)

#Here we define an initial profile that isn't invariant upon propagation.
ψ₀ = lg(rs,rs) / 2 + lg(rs,rs,p=1)

ψ = free_propagation(ψ₀,rs,rs,1) #Then, we propagate it by a distance z=1.

#Here are the initial profiles the propagated beam, side by side.
visualize([ψ₀,ψ] |> stack,ratio=2)

#We can also provide a collection of z values to produce an animation:
using StructuredLight

rs = LinRange(-12,12,256)
zs = LinRange(0,1,64)

#This is a laguerre-gaussian modulated by a sine function.
ψ₀ = lg(rs,rs,l=2,p=1) .* map(r->sin(6*r[2]),Iterators.product(rs,rs))

#Now the propagation is performed for each z ∈ zs. The output is a 3D array.
ψs = free_propagation(ψ₀,rs,rs,zs)

show_animation(ψs,ratio=2)

using StructuredLight

rs = LinRange(-4,4,256)
zs = LinRange(.01,1/2,32)

#This is a gaussian mode
ψ₀ = lg(rs,rs)
scalings = @. √(1+4*zs^2) #Here, we introduce the scalings given by w(z)/w0

#Now we propagate, including the scalings
ψs = free_propagation(ψ₀,rs,rs,zs,scalings)

#Note that the scalings compensate the diffraction of the beam.
#Therefore, the animation seems still.
show_animation(ψs,ratio=2)

References

The book "Schmidt, J. D. (2010). Numerical Simulation of Optical Wave Propagation with Examples in MATLAB. United States: SPIE." is a great resource to learn about numerical propagation of paraxial light beams

kerr_propagation

StructuredLight.kerr_propagation — Function

kerr_propagation(ψ₀,xs,ys,zs,total_steps;k=1,g=1)

Solve ∇² ψ + 2ik ∂_z ψ = - g |ψ|² ψ under the initial condition ψ₀.

xs and ys are the grids over which ψ₀ is calculated.

The outputs are saved at every zs, which is a number or a collection of numbers.

total_steps is the number of steps over which we discretize the propagation. The larger the total_steps, the better the precision and the slower is the calculation.

Example

using StructuredLight

rs = LinRange(-2.5,2.5,256) #The transverse grid
zs = LinRange(0,.1,32) #The z grid

ψ₀ = lg(rs,rs) #Calculates the fundamental Laguerre-Gaussian mode

#We perform the propagation with a strong nonlinearity
ψ = kerr_propagation(ψ₀,rs,rs,zs,512,g=100)

show_animation(ψ,ratio=2) #The beam colapses due to the self focusing effect

References

The package NonlinearSchrodinger.jl has more available solvers for this equation, but, as far as I can see, it only works with one spatial dimensional. Its author has also written a paper that explains the theory that goes behind the numerical solution.

CUDA support

Both free_propagation and kerr_propagation can be run on Nvidia GPUs, which will greatly improve the performance of these functions. If you have one, you simply need to convert your initial profile to a CuArray and pass this converted array to the propagation methods (check the CUDA.jl documentation for more details). Then, multiple dispatch will do its magic!

Here is an example:

using StructuredLight
using CUDA #It is necessary to load the CUDA package

ψ₀ = lg(rs,rs) |> cu #Transfers array to GPU

ψ = free_propagation(ψ₀,rs,rs,zs) #This is running on the GPU!